// src/app/shared/models/form-name.ts
export class FormNameCls {
    patient: number;
    riskAssessment: number;
    sharedPathway: number;
    recoveryStar: number;
    progressNote: number;
    moveOnFunctional: number;
    patientTransaction: number;
    form13: number;
    form15: number;
    form17: number;
    // locationforms
    form10: number;
    form12: number;
    form14: number;
    form16: number;
    form2: number;
    form7: number;
    form8: number;
    form9: number;
    form11: number;
    form19: number;
    form20: number;
    form21: number;
    constructor() {
        this.patient = 0;
        this.riskAssessment = 0;
        this.sharedPathway = 0;
        this.recoveryStar = 0;
        this.progressNote = 0;
        this.moveOnFunctional = 0;
        this.patientTransaction = 0;
        this.form13 = 0;
        this.form15 = 0;
        this.form17 = 0;
        // locationforms
        this.form2 = 0;
        this.form7 = 0;
        this.form8 = 0;
        this.form9 = 0;
        this.form10 = 0;
        this.form11 = 0;
        this.form12 = 0;
        this.form14 = 0;
        this.form16 = 0;
        this.form19 = 0;
        this.form20 = 0;
        this.form21 = 0;
    }
}